moved setM5 to testPhotometry.py, where it is used to effectively validate the calcSNR_gamma method against the calcSNR_psf method
