+++ tests/testPhotometry.py
@@ -13,13 +13,75 @@ from lsst.sims.photUtils.Bandpass import Bandpass
 from lsst.sims.photUtils.Sed import Sed
 from lsst.sims.photUtils.EBV import EBVbase
 from lsst.sims.photUtils import PhotometryStars, PhotometryGalaxies, PhotometryBase, PhotometryHardware
-from lsst.sims.photUtils import LSSTdefaults, PhotometricParameters, setM5, calcSNR_gamma, calcGamma, \
-                                calcM5, calcSNR_psf
+from lsst.sims.photUtils import LSSTdefaults, PhotometricParameters, calcSNR_gamma, calcGamma, \
+                                calcM5, calcSNR_psf, expectedSkyCountsForM5
 from lsst.sims.photUtils.utils import testDefaults, cartoonPhotometryStars, \
                                       cartoonPhotometryGalaxies, testCatalog, cartoonStars, \
                                       cartoonGalaxies, testStars, testGalaxies, \
                                       cartoonStarsOnlyI, cartoonStarsIZ, \
                                       cartoonGalaxiesIG, galaxiesWithHoles
+
+
+def setM5(m5target, skysed, totalBandpass, hardware,
+          photParams,
+          seeing = None):
+    """
+    Take an SED representing the sky and normalize it so that
+    m5 (the magnitude at which an object is detected in this
+    bandpass at 5-sigma) is set to some specified value.
+
+    The 5-sigma limiting magnitude (m5) for an observation is
+    determined by a combination of the telescope and camera parameters
+    (such as diameter of the mirrors and the readnoise) together with the
+    sky background. This method (setM5) scales a provided sky background
+    Sed so that an observation would have a target m5 value, for the
+    provided hardware parameters. Using the resulting Sed in the
+    'calcM5' method will return this target value for m5.
+
+    @param [in] the desired value of m5
+
+    @param [in] skysed is an instantiation of the Sed class representing
+    sky emission
+
+    @param [in] totalBandpass is an instantiation of the Bandpass class
+    representing the total throughput of the telescope (instrumentation
+    plus atmosphere)
+
+    @param [in] hardware is an instantiation of the Bandpass class representing
+    the throughput due solely to instrumentation.
+
+    @param [in] photParams is an instantiation of the
+    PhotometricParameters class that carries details about the
+    photometric response of the telescope.
+
+    @param [in] seeing in arcseconds
+
+    @param [out] returns an instantiation of the Sed class that is the skysed renormalized
+    so that m5 has the desired value.
+
+    Note that the returned SED will be renormalized such that calling the method
+    self.calcADU(hardwareBandpass) on it will yield the number of counts per square
+    arcsecond in a given bandpass.
+    """
+
+    #This is based on the LSST SNR document (v1.2, May 2010)
+    #www.astro.washington.edu/users/ivezic/Astr511/LSST_SNRdoc.pdf
+
+    if seeing is None:
+        seeing = LSSTdefaults().seeing('r')
+
+    skyCountsTarget = expectedSkyCountsForM5(m5target, totalBandpass, seeing=seeing,
+                                             photParams=photParams)
+
+    skySedOut = Sed(wavelen=numpy.copy(skysed.wavelen),
+                    flambda=numpy.copy(skysed.flambda))
+
+    skyCounts = skySedOut.calcADU(hardware, photParams=photParams) \
+                    * photParams.platescale * photParams.platescale
+    skySedOut.multiplyFluxNorm(skyCountsTarget/skyCounts)
+
+    return skySedOut
+
 
 class variabilityUnitTest(unittest.TestCase):
